Divide and conquer
