+++ tests/system/README.md
@@ -1,0 +1,5 @@
+# System tests
+
+These tests are system level and depend on the current code structure
+and go toolchain behaviour. They are flaky as code and the go toolchain
+changes and drifts from what's asserted here.
